==> 01.1 SFPD Data Tests.ipynb <==
import pandas as pd
from datetime import date, datetime, timedelta

df = pd.read_csv('SFPD_preprocessed.csv')
neighbourhoods_df = pd.read_csv('SF_neighborhoods.csv')

# intended values for dataset
expected_neighborhoods = sorted(neighbourhoods_df['name'].tolist())
expected_start_day = 1
expected_start_month = 7
expected_start_year = 2019
expected_start_date = date (expected_start_year,expected_start_month,expected_start_day)
expected_end_day = 11
expected_end_month = 11
expected_end_year = 2020
expected_end_date = date (expected_end_year,expected_end_month,expected_end_day)
num_neighbourhoods = 117
date_delta = (expected_end_date - expected_start_date).days + 1
expected_num_records = num_neighbourhoods * date_delta
days = ["Monday","Tuesday","Wednesday","Thursday","Friday","Saturday","Sunday"]
expected_columns = ['Incident Date','Incident Year','Neighborhood','Incident Day of Week','Todays Reports','Reports 1 day ago',
                    'Reports 2 days ago','Reports 3 days ago','Reports 4 days ago','Reports 5 days ago','Reports 6 days ago',
                    'Reports 7 days ago','Reports 14 days ago','Reports 30 days ago','Reports 365 days ago',
                    'Last 7 days reports','Last 14 days reports','Last 28 days reports']

df['Incident Date'] = pd.to_datetime(df['Incident Date'])
first_date = df['Incident Date'].iloc[0]
last_date = df['Incident Date'].iloc[-1]

# assert intended values are as expected
assert expected_start_date == first_date, "First date is " + str(first_date) + " but should be " + str(expected_start_date)
print("First date of dataset is as expected.")
assert expected_end_date == last_date, "First date is " + str(last_date) + " but should be " + str(expected_last_date)
print("Last date of dataset is as expected.")
assert expected_num_records == len(df), "Expected " + str(expected_num_records) + " records but got " + str(len(df))
print("Number of records in dataset is as expected.")
actual_columns = df.columns
for i in range(len(expected_columns)):
    assert expected_columns[i] == actual_columns[i], "Expected column" + expected_columns[i] + " but got " + actual_columns[i]
print("Column names of dataset are as expected.")
current_expected_date = expected_start_date
neighbourhood_count = 1
for record_num in range(0,len(df.index)):
    record_date = df['Incident Date'].iloc[record_num]
    assert current_expected_date == record_date, ("Expected date " + str(current_expected_date) + " on row " + 
                                                  str(record_num) + " but got " + str(record_date))
    expected_neighbourhood = expected_neighborhoods[neighbourhood_count-1]
    actual_neighbourhood = df['Neighborhood'].iloc[record_num]
    assert expected_neighbourhood == actual_neighbourhood, ("Expected neighborhood " + expected_neighbourhood + " on row " +
                                                            str(record_num) + " but got " + actual_neighbourhood)
    expected_year = current_expected_date.year
    actual_year = df['Incident Year'].iloc[record_num]
    assert expected_year == actual_year, ("Expected year " + expected_year + " on row " + str(record_num) + " but got " + 
                                          actual_year)
    expected_day_of_week = days[current_expected_date.weekday()]
    actual_day_of_week = df['Incident Day of Week'].iloc[record_num]
    assert expected_day_of_week == actual_day_of_week, ("Expected " + expected_day_of_week + " on row " + str(record_num) + 
                                                        " but got " + actual_day_of_week)
    j = 4
    while j < len(expected_columns):
        assert df[expected_columns[j]].iloc[record_num] >= 0, "Value for " + expected_columns[j] + " is less than 0"
        j += 1
    if neighbourhood_count == num_neighbourhoods:
        current_expected_date += timedelta(days=1)
        neighbourhood_count = 0
    neighbourhood_count += 1
print("All dates, years, neighbourhoods, days of week and values in dataset are as expected.")
print("\nAll tests completed successfully.")
