==> 04 - Exploratory Data Analysis.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv('finalised_data.csv')

df.describe()

Compare all columns of data against all other columns to visualise relationships

sns.pairplot(df, height=2.5);
plt.show()

Correlation analysis of features

pd.options.display.float_format = '{:,.2f}'.format
pd.set_option('display.max_columns', None)

df.corr()

Heatmap displaying correlation of features

plt.figure(figsize = (50,50))
sns.heatmap(df.corr(), annot = True)
plt.show()
