==> 05.1 - Split into train and test revised data.ipynb <==
import sklearn
from sklearn.model_selection import train_test_split
import pandas as pd

df = pd.read_csv('replica_finalised_data.csv')

y = df[['Todays Reports']]
x = df.drop('Todays Reports',axis=1)

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.25, random_state = 0)

y_train.shape, y_test.shape

x_train.shape, x_test.shape

y_train.to_csv('y_train_revised.csv', index = False)
x_train.to_csv('x_train_revised.csv', index = False)
y_test.to_csv('y_test_revised.csv', index = False)
x_test.to_csv('x_test_revised.csv', index = False)

# Sample sets also created for initial hyperparameter exploration
#This was necessary due to time and resource constraints and processing the entire training set thousands of times
#per algorithm only to obtain a rough feel for the data was not feasible (the time difference is exponential)
#Google Colab was explored as a solution however it was also limited in the processing power it offered
#The full training dataset will still be used for final hyperparameter training and cross-fold validation

x_sample, x_remainder, y_sample, y_remainder = train_test_split(x_train, y_train, test_size = 0.75, random_state = 0)

y_sample.shape, y_remainder.shape

x_sample.shape, x_remainder.shape

y_sample.to_csv('y_train_sample_revised.csv', index = False)
x_sample.to_csv('x_train_sample_revised.csv', index = False)
