==> 05.2 - Split into test and train tests.ipynb <==
import pandas as pd
import numpy as np

# load all files to be tested
y_train = pd.read_csv('y_train.csv')
x_train = pd.read_csv('x_train.csv')
y_test = pd.read_csv('y_test.csv')
x_test = pd.read_csv('x_test.csv')
y_train_sample = pd.read_csv('y_train_sample.csv')
x_train_sample = pd.read_csv('x_train_sample.csv')
y_train_revised = pd.read_csv('y_train_revised.csv')
x_train_revised = pd.read_csv('x_train_revised.csv')
y_test_revised = pd.read_csv('y_test_revised.csv')
x_test_revised = pd.read_csv('x_test_revised.csv')
y_train_sample_revised = pd.read_csv('y_train_sample_revised.csv')
x_train_sample_revised = pd.read_csv('x_train_sample_revised.csv')

# expected values
train_expected_length = 43875
test_expected_length = 14625
sample_expected_length = 10968
expected_x_columns = ['Reports 1 day ago','Reports 2 days ago','Reports 3 days ago','Reports 4 days ago','Reports 5 days ago',
                    'Reports 6 days ago','Reports 7 days ago','Reports 14 days ago','Reports 30 days ago',
                    'Reports 365 days ago','Last 7 days reports','Last 14 days reports','Last 28 days reports',
                    'Number of businesses','Businesses 1 day ago','Businesses 2 days ago','Businesses 3 days ago',
                    'Businesses 4 days ago','Businesses 5 days ago','Businesses 6 days ago','Businesses 7 days ago',
                    'Businesses 14 days ago','Businesses 30 days ago','Businesses 365 days ago','Number of closures',
                    'Closures 1 day ago','Closures 2 days ago','Closures 3 days ago','Closures 4 days ago',
                    'Closures 5 days ago','Closures 6 days ago','Closures 7 days ago','Closures 14 days ago',
                    'Closures 30 days ago','Closures 365 days ago','Last 7 days closures','Last 14 days closures',
                    'Last 28 days closures','Number of openings','Openings 1 day ago','Openings 2 days ago',
                    'Openings 3 days ago','Openings 4 days ago','Openings 5 days ago','Openings 6 days ago',
                    'Openings 7 days ago','Openings 14 days ago','Openings 30 days ago','Openings 365 days ago',
                    'Last 7 days openings','Last 14 days openings','Last 28 days openings','Alamo Square','Anza Vista',
                    'Apparel City','Aquatic Park / Ft. Mason','Balboa Terrace','Bayview','Bernal Heights','Bret Harte',
                    'Buena Vista','Candlestick Point SRA','Castro','Cathedral Hill','Cayuga','Central Waterfront','Chinatown',
                    'Civic Center','Clarendon Heights','Cole Valley','Corona Heights','Cow Hollow','Crocker Amazon',
                    'Diamond Heights','Dogpatch','Dolores Heights','Downtown / Union Square','Duboce Triangle','Eureka Valley',
                    'Excelsior','Fairmount','Financial District','Fishermans Wharf','Forest Hill','Forest Knolls','Glen Park',
                    'Golden Gate Heights','Golden Gate Park','Haight Ashbury','Hayes Valley','Holly Park','Hunters Point',
                    'India Basin','Ingleside','Ingleside Terraces','Inner Richmond','Inner Sunset','Japantown','Laguna Honda',
                    'Lake Street','Lakeshore','Laurel Heights / Jordan Park','Lincoln Park / Ft. Miley','Little Hollywood',
                    'Lone Mountain','Lower Haight','Lower Nob Hill','Lower Pacific Heights','Marina','McLaren Park',
                    'Merced Heights','Merced Manor','Midtown Terrace','Mint Hill','Miraloma Park','Mission','Mission Bay',
                    'Mission Dolores','Mission Terrace','Monterey Heights','Mt. Davidson Manor','Nob Hill','Noe Valley',
                    'North Beach','Northern Waterfront','Oceanview','Outer Mission','Outer Richmond','Outer Sunset',
                    'Pacific Heights','Panhandle','Parkmerced','Parkside','Parnassus Heights','Peralta Heights',
                    'Polk Gulch','Portola','Potrero Hill','Presidio Heights','Presidio National Park','Presidio Terrace',
                    'Produce Market','Rincon Hill','Russian Hill','Seacliff','Sherwood Forest','Showplace Square',
                    'Silver Terrace','South Beach','South of Market','St. Francis Wood','St. Marys Park','Stonestown',
                    'Sunnydale','Sunnyside','Sutro Heights','Telegraph Hill','Tenderloin','Treasure Island','Union Street',
                    'University Mound','Upper Market','Visitacion Valley','West Portal','Western Addition',
                    'Westwood Highlands','Westwood Park','Yerba Buena Island','Friday','Saturday','Sunday','Thursday',
                    'Tuesday','Wednesday']
expected_y_columns = ['Todays Reports']

# method for testing datasets
def test_dataset(df, expected_length, expected_columns):
    assert len(df) == expected_length, "Dataset length not as expected"
    for col in df.columns:
        assert col in expected_columns, "Column " + str(col) + " in df but not expected."
    for col in expected_columns:
        assert col in df.columns, "Column " + str(col) + " expected but not found"
    for record_num in range(0,len(df.index)):
        j = 0
        while j < len(expected_columns):
            assert df[expected_columns[j]].iloc[record_num] != np.nan, "no value for " + expected_columns[j]
            j += 1
    print("All tests passed successfully.")

<h4> Testing starts here <h4>

test_dataset(y_train, train_expected_length, expected_y_columns)

test_dataset(x_train, train_expected_length, expected_x_columns)

test_dataset(y_test, test_expected_length, expected_y_columns)

test_dataset(x_test, test_expected_length, expected_x_columns)

test_dataset(y_train_sample, sample_expected_length, expected_y_columns)

test_dataset(x_train_sample, sample_expected_length, expected_x_columns)

#revised dataset retained Date and Neighborhood for testing and manipulation purposes
expected_x_columns.append('Date')
expected_x_columns.append('Neighborhood')

test_dataset(y_train_revised, train_expected_length, expected_y_columns)

test_dataset(x_train_revised, train_expected_length, expected_x_columns)

test_dataset(y_test_revised, test_expected_length, expected_y_columns)

test_dataset(x_test_revised, test_expected_length, expected_x_columns)

test_dataset(y_train_sample_revised, sample_expected_length, expected_y_columns)

test_dataset(x_train_sample_revised, sample_expected_length, expected_x_columns)
