==> 06 - Create GUI dataset.ipynb <==
import pandas as pd
import datetime
from datetime import date
import numpy as np

# Create dummy dataset for use with GUI

x = pd.read_csv('x_test.csv')
y = pd.read_csv('y_test.csv')

x.head()

df = pd.read_csv('SF_neighborhoods.csv')
neighbourhoods = []
num_neighbourhoods = len(df.index)
for i in range(0,num_neighbourhoods):
    neighbourhoods.append(df.iloc[i,2])

days = ['Friday','Saturday','Sunday','Thursday','Tuesday','Wednesday']

x

y

# merge x and y data as both required for GUI
merged = pd.concat([x,y], axis = 1, join = "inner")
merged

# generate recent dates for dataset
num_neighbourhoods = len(neighbourhoods)
num_records = len(merged.index)
num_dates = num_records/num_neighbourhoods
num_dates

current_date = date.today()
current_date_array = []
i = 0
while i < num_neighbourhoods:
    current_date_array.append(current_date)
    i+=1

dates = pd.DataFrame()
final_dates = pd.DataFrame()
dates['Date'] = current_date_array
final_dates['Date'] = []
i = 1
while i < num_dates:
    final_dates = final_dates.append(dates)
    current_date += datetime.timedelta(days=1)
    dates['Date'] = current_date
    i+=1
final_dates.reset_index(drop=True, inplace=True)
final_dates

# add new dates to gui dataset
merged = pd.concat([merged,final_dates], axis = 1, join = "inner")
merged

# generate new neighbourhoods for dataset in balanced quantity
neighbourhood_data = pd.DataFrame()
neighbourhood_data['Neighborhood'] = neighbourhoods
final_neighbourhoods = pd.DataFrame()
final_neighbourhoods['Neighborhood'] = neighbourhoods
i = 1
while i < num_dates:
    final_neighbourhoods = final_neighbourhoods.append(neighbourhood_data)
    i+=1
final_neighbourhoods.reset_index(drop=True, inplace=True)
final_neighbourhoods

# add neighbourhoods to GUI dataset
merged = pd.concat([merged,final_neighbourhoods], axis = 1, join = "inner")
merged

merged.to_csv('data.csv', index = False)
