==> 07 - Feature Selection - Remove zero variance features (testing included).ipynb <==
import sklearn
import pandas as pd

x_train = pd.read_csv('x_train.csv')
x_train_original_length = len(x_train)

x_train.head()

<h3>Remove features with zero variance<h3>

from sklearn.feature_selection import VarianceThreshold
threshold = 0.0;
variance_threshold = VarianceThreshold(threshold = threshold) #.0005


#Get columns with zero variance
def get_zero_vairance_columns(df):
    try:
        variance_threshold.fit(df)
        zero_var_columns = [col for col in df
                            if col not in df.columns[variance_threshold.get_support()]]
    except:
        zero_var_columns = df.columns
    print(len(zero_var_columns),"features have",threshold,"variance.")
    return zero_var_columns

zero_var_columns = get_zero_vairance_columns(x_train)

#Drop all columns with threshold variance
def drop_zero_variance_columns(df,zero_var_columns):
    df = df.drop(zero_var_columns[:-1], axis = 1)
    return df

x_train = drop_zero_variance_columns(x_train, zero_var_columns)

x_train

x_train.to_csv('x_train_zero_variance_removed.csv', index = False)

<h3> Testing begins here <h3>

# assert zero variance removal works as intended
def test_zero_variance_check(filename, expected_remaining_columns):
    df = pd.read_csv(filename)
    original_columns = df.columns.tolist()
    expected_zero_var_columns = [x for x in original_columns if x not in expected_remaining_columns]
    zero_var_columns = get_zero_vairance_columns(df)
    assert len(zero_var_columns) == len(expected_zero_var_columns) +1 or (len(zero_var_columns) == 0 and len(expected_zero_var_columns) == 0), "Length of zero variance columns is different to expected"
    for col in zero_var_columns[:-1]:
        assert col in expected_zero_var_columns, "Zero var columns not as expected." + str(col)
    for col in expected_zero_var_columns:
        assert col in zero_var_columns, "Zero var columns not as expected."
    new_df = drop_zero_variance_columns(df,zero_var_columns)
    assert len(df) == len(new_df), "Dataframe lengths vary pre and post drop columns."
    new_df_columns = new_df.columns.tolist()
    assert len(new_df_columns) == len(expected_remaining_columns), "Post column drop datafram length not as expected"
    for col in new_df_columns:
        assert col in expected_remaining_columns, "Remaining columns not as expected."
    for col in expected_remaining_columns:
        assert col in new_df_columns, "Zero var columns not as expected."
    print("All tests completed successfully")

# test dataset where all features have zero variance
filename = 'x_train_all_zero_variance_test.csv'
expected_remaining_columns = ['Wednesday']
test_zero_variance_check(filename, expected_remaining_columns)

# test dataset where all but one feature has zero variance
filename = 'x_train_all_but_reports_one_day_ago_zero_variance_test.csv'
expected_remaining_columns = ['Reports 1 day ago','Wednesday']
test_zero_variance_check(filename, expected_remaining_columns)

# test dataset with no zero variance
filename = 'x_train.csv'
expected_remaining_columns = ['Reports 1 day ago','Reports 2 days ago','Reports 3 days ago','Reports 4 days ago','Reports 5 days ago',
                    'Reports 6 days ago','Reports 7 days ago','Reports 14 days ago','Reports 30 days ago',
                    'Reports 365 days ago','Last 7 days reports','Last 14 days reports','Last 28 days reports',
                    'Number of businesses','Businesses 1 day ago','Businesses 2 days ago','Businesses 3 days ago',
                    'Businesses 4 days ago','Businesses 5 days ago','Businesses 6 days ago','Businesses 7 days ago',
                    'Businesses 14 days ago','Businesses 30 days ago','Businesses 365 days ago','Number of closures',
                    'Closures 1 day ago','Closures 2 days ago','Closures 3 days ago','Closures 4 days ago',
                    'Closures 5 days ago','Closures 6 days ago','Closures 7 days ago','Closures 14 days ago',
                    'Closures 30 days ago','Closures 365 days ago','Last 7 days closures','Last 14 days closures',
                    'Last 28 days closures','Number of openings','Openings 1 day ago','Openings 2 days ago',
                    'Openings 3 days ago','Openings 4 days ago','Openings 5 days ago','Openings 6 days ago',
                    'Openings 7 days ago','Openings 14 days ago','Openings 30 days ago','Openings 365 days ago',
                    'Last 7 days openings','Last 14 days openings','Last 28 days openings','Alamo Square','Anza Vista',
                    'Apparel City','Aquatic Park / Ft. Mason','Balboa Terrace','Bayview','Bernal Heights','Bret Harte',
                    'Buena Vista','Candlestick Point SRA','Castro','Cathedral Hill','Cayuga','Central Waterfront','Chinatown',
                    'Civic Center','Clarendon Heights','Cole Valley','Corona Heights','Cow Hollow','Crocker Amazon',
                    'Diamond Heights','Dogpatch','Dolores Heights','Downtown / Union Square','Duboce Triangle','Eureka Valley',
                    'Excelsior','Fairmount','Financial District','Fishermans Wharf','Forest Hill','Forest Knolls','Glen Park',
                    'Golden Gate Heights','Golden Gate Park','Haight Ashbury','Hayes Valley','Holly Park','Hunters Point',
                    'India Basin','Ingleside','Ingleside Terraces','Inner Richmond','Inner Sunset','Japantown','Laguna Honda',
                    'Lake Street','Lakeshore','Laurel Heights / Jordan Park','Lincoln Park / Ft. Miley','Little Hollywood',
                    'Lone Mountain','Lower Haight','Lower Nob Hill','Lower Pacific Heights','Marina','McLaren Park',
                    'Merced Heights','Merced Manor','Midtown Terrace','Mint Hill','Miraloma Park','Mission','Mission Bay',
                    'Mission Dolores','Mission Terrace','Monterey Heights','Mt. Davidson Manor','Nob Hill','Noe Valley',
                    'North Beach','Northern Waterfront','Oceanview','Outer Mission','Outer Richmond','Outer Sunset',
                    'Pacific Heights','Panhandle','Parkmerced','Parkside','Parnassus Heights','Peralta Heights',
                    'Polk Gulch','Portola','Potrero Hill','Presidio Heights','Presidio National Park','Presidio Terrace',
                    'Produce Market','Rincon Hill','Russian Hill','Seacliff','Sherwood Forest','Showplace Square',
                    'Silver Terrace','South Beach','South of Market','St. Francis Wood','St. Marys Park','Stonestown',
                    'Sunnydale','Sunnyside','Sutro Heights','Telegraph Hill','Tenderloin','Treasure Island','Union Street',
                    'University Mound','Upper Market','Visitacion Valley','West Portal','Western Addition',
                    'Westwood Highlands','Westwood Park','Yerba Buena Island','Friday','Saturday','Sunday','Thursday',
                    'Tuesday','Wednesday']
test_zero_variance_check(filename, expected_remaining_columns)

#test final file created and verify all columns remain
filename = 'x_train_zero_variance_removed.csv'
expected_remaining_columns = ['Reports 1 day ago','Reports 2 days ago','Reports 3 days ago','Reports 4 days ago','Reports 5 days ago',
                    'Reports 6 days ago','Reports 7 days ago','Reports 14 days ago','Reports 30 days ago',
                    'Reports 365 days ago','Last 7 days reports','Last 14 days reports','Last 28 days reports',
                    'Number of businesses','Businesses 1 day ago','Businesses 2 days ago','Businesses 3 days ago',
                    'Businesses 4 days ago','Businesses 5 days ago','Businesses 6 days ago','Businesses 7 days ago',
                    'Businesses 14 days ago','Businesses 30 days ago','Businesses 365 days ago','Number of closures',
                    'Closures 1 day ago','Closures 2 days ago','Closures 3 days ago','Closures 4 days ago',
                    'Closures 5 days ago','Closures 6 days ago','Closures 7 days ago','Closures 14 days ago',
                    'Closures 30 days ago','Closures 365 days ago','Last 7 days closures','Last 14 days closures',
                    'Last 28 days closures','Number of openings','Openings 1 day ago','Openings 2 days ago',
                    'Openings 3 days ago','Openings 4 days ago','Openings 5 days ago','Openings 6 days ago',
                    'Openings 7 days ago','Openings 14 days ago','Openings 30 days ago','Openings 365 days ago',
                    'Last 7 days openings','Last 14 days openings','Last 28 days openings','Alamo Square','Anza Vista',
                    'Apparel City','Aquatic Park / Ft. Mason','Balboa Terrace','Bayview','Bernal Heights','Bret Harte',
                    'Buena Vista','Candlestick Point SRA','Castro','Cathedral Hill','Cayuga','Central Waterfront','Chinatown',
                    'Civic Center','Clarendon Heights','Cole Valley','Corona Heights','Cow Hollow','Crocker Amazon',
                    'Diamond Heights','Dogpatch','Dolores Heights','Downtown / Union Square','Duboce Triangle','Eureka Valley',
                    'Excelsior','Fairmount','Financial District','Fishermans Wharf','Forest Hill','Forest Knolls','Glen Park',
                    'Golden Gate Heights','Golden Gate Park','Haight Ashbury','Hayes Valley','Holly Park','Hunters Point',
                    'India Basin','Ingleside','Ingleside Terraces','Inner Richmond','Inner Sunset','Japantown','Laguna Honda',
                    'Lake Street','Lakeshore','Laurel Heights / Jordan Park','Lincoln Park / Ft. Miley','Little Hollywood',
                    'Lone Mountain','Lower Haight','Lower Nob Hill','Lower Pacific Heights','Marina','McLaren Park',
                    'Merced Heights','Merced Manor','Midtown Terrace','Mint Hill','Miraloma Park','Mission','Mission Bay',
                    'Mission Dolores','Mission Terrace','Monterey Heights','Mt. Davidson Manor','Nob Hill','Noe Valley',
                    'North Beach','Northern Waterfront','Oceanview','Outer Mission','Outer Richmond','Outer Sunset',
                    'Pacific Heights','Panhandle','Parkmerced','Parkside','Parnassus Heights','Peralta Heights',
                    'Polk Gulch','Portola','Potrero Hill','Presidio Heights','Presidio National Park','Presidio Terrace',
                    'Produce Market','Rincon Hill','Russian Hill','Seacliff','Sherwood Forest','Showplace Square',
                    'Silver Terrace','South Beach','South of Market','St. Francis Wood','St. Marys Park','Stonestown',
                    'Sunnydale','Sunnyside','Sutro Heights','Telegraph Hill','Tenderloin','Treasure Island','Union Street',
                    'University Mound','Upper Market','Visitacion Valley','West Portal','Western Addition',
                    'Westwood Highlands','Westwood Park','Yerba Buena Island','Friday','Saturday','Sunday','Thursday',
                    'Tuesday','Wednesday']
test_zero_variance_check(filename, expected_remaining_columns)

final_df = pd.read_csv(filename)
assert x_train_original_length == len(final_df), "Final datafile has an unexpected number of rows."
print("Final dataset is of the expected length.")
